# FoodVerse-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/food-bags": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new food bag for a store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food-bags"
                ],
                "summary": "Create a new food bag",
                "parameters": [
                    {
                        "description": "Food bag data",
                        "name": "foodbag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FoodBagInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerFoodBag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/food-bags/search": {
            "post": {
                "description": "Search for food bags near a location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food-bags"
                ],
                "summary": "Search nearby food bags",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerFoodBagSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SwaggerFoodBagResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/food-bags/{id}": {
            "get": {
                "description": "Get food bag details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food-bags"
                ],
                "summary": "Get food bag by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food bag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerFoodBag"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update food bag details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food-bags"
                ],
                "summary": "Update food bag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food bag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Food bag data",
                        "name": "foodbag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FoodBagInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerFoodBag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a food bag",
                "tags": [
                    "food-bags"
                ],
                "summary": "Delete food bag",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food bag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticate user and return JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Register a new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stores/{store_id}/food-bags": {
            "get": {
                "description": "Get all food bags for a specific store",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "food-bags"
                ],
                "summary": "Get food bags by store",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SwaggerFoodBag"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get current user profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SwaggerUser"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AuthResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2023-01-02T00:00:00Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/model.SwaggerUser"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "model.FoodBagInput": {
            "type": "object",
            "required": [
                "discounted_price",
                "original_price",
                "pickup_time_end",
                "pickup_time_start",
                "quantity_total",
                "store_id",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discounted_price": {
                    "type": "number"
                },
                "image_url": {
                    "type": "string"
                },
                "original_price": {
                    "type": "number"
                },
                "pickup_time_end": {
                    "type": "string"
                },
                "pickup_time_start": {
                    "type": "string"
                },
                "quantity_total": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "model.SwaggerFoodBag": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Mixed"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Assorted fresh items"
                },
                "discounted_price": {
                    "type": "number",
                    "example": 10
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "original_price": {
                    "type": "number",
                    "example": 25
                },
                "pickup_time_end": {
                    "type": "string",
                    "example": "2023-01-01T21:00:00Z"
                },
                "pickup_time_start": {
                    "type": "string",
                    "example": "2023-01-01T18:00:00Z"
                },
                "quantity_left": {
                    "type": "integer",
                    "example": 3
                },
                "quantity_total": {
                    "type": "integer",
                    "example": 5
                },
                "store": {
                    "$ref": "#/definitions/model.SwaggerStore"
                },
                "store_id": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Mixed Food Bag"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "model.SwaggerFoodBagResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Mixed"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Assorted fresh items"
                },
                "discount_percent": {
                    "type": "integer",
                    "example": 60
                },
                "discounted_price": {
                    "type": "number",
                    "example": 10
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "original_price": {
                    "type": "number",
                    "example": 25
                },
                "pickup_time_end": {
                    "type": "string",
                    "example": "2023-01-01T21:00:00Z"
                },
                "pickup_time_start": {
                    "type": "string",
                    "example": "2023-01-01T18:00:00Z"
                },
                "quantity_left": {
                    "type": "integer",
                    "example": 3
                },
                "store": {
                    "$ref": "#/definitions/model.SwaggerStoreResponse"
                },
                "title": {
                    "type": "string",
                    "example": "Mixed Food Bag"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "model.SwaggerFoodBagSearchRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Mixed"
                },
                "latitude": {
                    "type": "number",
                    "example": 40.7128
                },
                "longitude": {
                    "type": "number",
                    "example": -74.006
                },
                "max_price": {
                    "type": "number",
                    "example": 20
                },
                "min_price": {
                    "type": "number",
                    "example": 5
                },
                "radius": {
                    "type": "number",
                    "example": 5
                }
            }
        },
...
        "model.SwaggerStore": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "category": {
                    "type": "string",
                    "example": "Grocery"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Fresh organic produce"
                },
                "email": {
                    "type": "string",
                    "example": "store@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/store.jpg"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "latitude": {
                    "type": "number",
                    "example": 40.7128
                },
                "longitude": {
                    "type": "number",
                    "example": -74.006
                },
                "name": {
                    "type": "string",
                    "example": "Green Grocery"
                },
                "owner_id": {
                    "type": "integer",
                    "example": 1
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "model.SwaggerUser": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "456 Oak St"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_type": {
                    "type": "string",
                    "example": "consumer"
                }
            }
        },
        "model.UserInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "user_type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "user_type": {
                    "$ref": "#/definitions/model.UserType"
                }
            }
        },
        "model.UserType": {
            "type": "string",
            "enum": [
                "consumer",
                "business"
            ],
            "x-enum-varnames": [
                "UserTypeConsumer",
                "UserTypeBusiness"
            ]
        }
    }
}